Actor refactoring - V1
The project has been refactored in order to use the actor model for all the application logic. This led to the creation of several actors managing the various aspects of the scalachain node. One notable improvement is the use of Akka Persistence actor to store snapshots of the blockchain every time a new block is mined, so when the node is restarted again the status is restored according to the snapshot.