Development (#8)
* added travis support

* added travis script for tests

* added Travis.ci status badge

* removed useless main file

* fix Transaction json converter

* Docker integration (#3)

* Docker integration

* added dockerhub link

* fix init.sh
init.sh has been fixed adding the git fetch step
before the checkout of the branch.

improvement docker/README.md
added more information in the docker readme file

* Actor refactoring - V1

The project has been refactored in order to use the actor model for all the application logic. This led to the creation of several actors managing the various aspects of the scalachain node. One notable improvement is the use of Akka Persistence actor to store snapshots of the blockchain every time a new block is mined, so when the node is restarted again the status is restored according to the snapshot.

* update mining process

Miner actor has been updated (ad weel as the entire node) to better
handle concurrency in mining operations.

* fix pow test

pow test has been fixed to take into account the new validity
condition for proofs.

* refactoring blockchain test

blockchain tests have been refactored to exploit flatspec style of testing

* removed old tests

* - improvement: actors tests
- refactoring: project structure

The "low level" actors (i.e. Broker, Miner, Blockchain) have been tested
to check if they correctly provide the required functionalities.

The project structure has been refactored in order to create a
correct structure of packages.

* - fix refactoring error

fixed a minor "error" due to refactoring

* Added coinbase transaction

Coinbase transaction (reward for the mining of a new block) was missing.
When a new block is mined, a coinbase transaction with value 100 is
registered with recipient the node id.

* Documentation (#7)

* documentation v1

* documentation v2

Added a paragraph about the longest chain rule.
Added the image about the Scalachain node and the actor hierarchy.
Small grammar corrections