Docker improvement
Improved docker image creation. Sbt is downloaded during image creation
to reduce start time of containers.
Source code can be mounted in a folderinside the container and then
copied in the /development one. This in order to prevent problems due to
concurrent r/w in the same volume by multiple containers.