Cluster (#19)
* Initial cluster integration

- added cluster configuration
- added cluster manager and listener
- added /status/members endpoint
- added docker-compose file
- improved application configuration

* Transaction broadcast first implementation

* Transaction broadcast refactoring

- transaction broadcast has been refactored.

* Docker improvement

Improved docker image creation. Sbt is downloaded during image creation
to reduce start time of containers.
Source code can be mounted in a folderinside the container and then
copied in the /development one. This in order to prevent problems due to
concurrent r/w in the same volume by multiple containers.

* Fixed tests

Tests are fixed to support the new cluster architecture

* New block broadcast

The broadcast of new blocks has been implemented.
Fixed the persistence of snapshots in containers.

* Updated documentation for the new cluster feature